+++ spec-integration/common.ts
@@ -4,5 +4,7 @@ import { config } from 'dotenv';
 export const setEnvs = () => {
   if (existsSync('.env')) {
     config();
+  } else {
+    throw new Error('.env doesn\'t exist!');
   }
 };