+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0 (May 06, 2022)
+* Now method `uploadAttachment` from `AttachmentProcessor` automatically sets the `contentType`
+* Fixed retrying logic for method `uploadAttachment`
+* Removed validation of the file size (`REQUEST_MAX_BODY_LENGTH` env) because it not works with streams 
+
 ## 2.0.2 (April 13, 2022)
 * Fix dependencies